backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
strawberry-graphql[fastapi]==0.224.0
SQLAlchemy==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
qdrant-client==1.8.0
redis==5.0.3
httpx==0.27.0
